refactor: myReviewAPI 위치 수정
